requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
matplotlib>=3.7.0
numpy>=1.24.0